+++ packages/react-widgets/src/widgets/HistogramWidget.js
@@ -56,7 +56,7 @@ function HistogramWidget(props) {
   const tooltipFormatter = useCallback(
     ([serie]) => {
       const formattedValue = formatter
-        ? formatter(serie.value)
+        ? formatter(serie.value, serie.dataIndex, ticks)
         : { prefix: '', value: serie.value };
 
       return `${
@@ -65,7 +65,7 @@ function HistogramWidget(props) {
           : formattedValue
       }`;
     },
-    [formatter]
+    [formatter, ticks]
   );
 
   useEffect(() => {